Create serialization subpackage
