+++ baybe/searchspace/discrete.py
@@ -19,8 +19,8 @@ from baybe.parameters import (
 )
 from baybe.parameters.base import DiscreteParameter, Parameter
 from baybe.searchspace.validation import validate_parameter_names
+from baybe.serialization import converter, select_constructor_hook
 from baybe.utils import df_drop_single_value_columns, eq_dataframe, fuzzy_row_match
-from baybe.utils.serialization import converter, select_constructor_hook
 
 _METADATA_COLUMNS = ["was_recommended", "was_measured", "dont_recommend"]
