+++ baybe/serialization/__init__.py
@@ -1,0 +1,21 @@
+"""Serialization functionality."""
+
+from baybe.serialization.core import (
+    block_deserialization_hook,
+    block_serialization_hook,
+    converter,
+    get_base_structure_hook,
+    select_constructor_hook,
+    unstructure_base,
+)
+from baybe.serialization.mixin import SerialMixin
+
+__all__ = [
+    "block_deserialization_hook",
+    "block_serialization_hook",
+    "converter",
+    "get_base_structure_hook",
+    "select_constructor_hook",
+    "unstructure_base",
+    "SerialMixin",
+]
